refactor: move the easing helpers out of vtbackend and into crispy
Animation.hpp holds LinearEasing, EaseOutCubic and AnimationState -- sixty lines
of easing arithmetic over <algorithm>, <cassert> and <chrono>, with nothing in it
that knows a terminal exists. It has one consumer, Terminal.hpp, which uses it for
the screen-transition and cursor-motion animations.

Groundwork for splitting vtbackend into concern directories: a general-purpose
utility with one consumer is exactly what makes a `core/` directory into a junk
drawer, and the rule the split is built on -- a header earns a place in core/ only
if two other directories or another module name it -- says this one does not.

crispy is where the tree already puts these. Same reasoning, same move as
AsciiText.h -> crispy/ascii.h in df037b9c, and ScopedTimer in 1b3ec7a2.
