+++ src/crispy/Animation_test.cpp
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0
-#include <vtbackend/Animation.hpp>
+#include <crispy/Animation.hpp>
 
 #include <catch2/catch_approx.hpp>
 #include <catch2/catch_test_macros.hpp>
 
-using namespace vtbackend;
+using namespace crispy;
 using namespace std::chrono_literals;
 using Catch::Approx;
